+++ base.php
@@ -2,7 +2,7 @@
 <body <?php body_class(); ?>>
 
   <!--[if lt IE 8]>
-    <div class="alert alert-warning">
+    <div class="alert-box warning">
       <?php _e('You are using an <strong>outdated</strong> browser. Please <a href="http://browsehappy.com/">upgrade your browser</a> to improve your experience.', 'forward'); ?>
     </div>
   <![endif]-->